Added legalRepReferenceNumber
